UPDATE Scirptables
Added:
	Interests Scriptable
	All Buildings
	Added mesh and image to people

FIx :
	Peoples scriptable getting reseted

+++ Assets/Script/Other/Building.cs
@@ -46,7 +46,7 @@ public class Building : ScriptableObject
 
 
             [SerializeField] private InterestCategory attribute;
-            [SerializeField] private int bonus;
+            [SerializeField] private float bonus;
             [SerializeField] private BonusType bonusType;
         }
     }
